Separate routingController and tests submit data
Problem was that a form-submit in the tests doesn't submits the data correctly.
To get this problem solved, the request-Object has to be parameter of the controller function. For this the index-route is separated now to indexAction for GET and indexPostAction for POST-Requests in router.php.
To apply the Crawler-Filter in a more consistant way, the form in '/' and the link in '/link/' has css-id's.