Remove Request constructor-param from Controller
As every controller-Action is mapped to a closure, it is more conveniant to add the Request-Parameter to every controller who uses this data, istead of passing it to the constructor.
So the protected $request property of the controller-class is removed.
The two functions indexPostAction and deleteAction are getting the Request injected.