Use doctrine-orm and eventhandler to execute crypt
The persistance layer works now with Doctrine ORM.
So an Entity Credentials, used by an Entitymanager is doing the persistance stuff in the controller.
The CredentialService is obsolete.
The encryption and decryption of the Credentioal-Object is event-driven by doctrine-postLoad an prePersist events.
The Table scheme can be generated by console-commands. The readme is adapted and the mysql-dump and installfolder is not needed anymore.